+++ scripts/sdxl_textual_inversion.py
@@ -12,7 +12,7 @@ from library.utils.device_utils import init_ipex
 from library.strategies import strategy_sdxl, strategy_sd
 from library.data.dataset import DatasetGroup, MinimalDataset
 from library.training.sdxl_sample_generation import sample_images
-from library.training.sdxl_model_prep import load_target_model as load_target_model_sdxl
+from library.models.sdxl_model_prep import load_target_model as load_target_model_sdxl
 from library.config.dataclasses.sdxl_textual_inversion import SDXLTextualInversionConfig
 from library.config.config_validation import prepare_config, validate_config, validate_sdxl_textual_inversion